# packages/contracts/openapi/api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/media": {
      "get": {
        "operationId": "MediaAdminController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "quarantined",
                "ready",
                "rejected"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          },
          {
            "name": "unused",
            "required": false,
            "in": "query",
            "description": "Only assets that are not used anywhere",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MediaAssetDto"
                  }
                }
              }
            }
          }
        },
        "summary": "List media assets (status, q, unused)",
        "tags": [
          "admin-media"
        ]
      }
    },
    "/api/v1/admin/media/uploads": {
      "post": {
        "operationId": "MediaAdminController_requestUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTicketDto"
                }
              }
            }
          }
        },
        "summary": "Request a short-lived signed upload for the private quarantine bucket",
        "tags": [
          "admin-media"
        ]
      }
    },
    "/api/v1/admin/media/{id}/complete": {
      "post": {
        "operationId": "MediaAdminController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAssetDto"
                }
              }
            }
          }
        },
        "summary": "Validate the uploaded bytes and queue variant processing",
        "tags": [
          "admin-media"
        ]
      }
    },
    "/api/v1/admin/media/{id}": {
      "get": {
        "operationId": "MediaAdminController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAssetDto"
                }
              }
            }
          }
        },
        "tags": [
          "admin-media"
        ]
      },
      "patch": {
        "operationId": "MediaAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMediaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAssetDto"
                }
              }
            }
          }
        },
        "summary": "Edit alt text, credit, rights note and focal point",
        "tags": [
          "admin-media"
        ]
      },
      "delete": {
        "operationId": "MediaAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete an unused asset; refused while any usage exists",
        "tags": [
          "admin-media"
        ]
      }
    },
    "/api/v1/admin/businesses/{businessId}/gallery": {
      "get": {
        "operationId": "BusinessGalleryController_get",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GalleryEntryDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin-businesses"
        ]
      },
      "put": {
        "operationId": "BusinessGalleryController_set",
        "parameters": [
          {
            "name": "businessId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetGalleryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GalleryEntryDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Replace the gallery (order, captions, alt overrides, cover) with expectedVersion",
        "tags": [
          "admin-businesses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Melbourne Sphere API",
    "description": "REST API for the Melbourne Sphere directory and blog. Admin routes require the ms_admin_session cookie.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ms_admin_session"
      }
    },
    "schemas": {
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "MediaVariantDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "thumbnail",
              "card",
              "hero"
            ]
          },
          "url": {
            "type": "string"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          }
        },
        "required": [
          "kind",
          "url",
          "width",
          "height"
        ]
      },
      "MediaUsageDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "business",
              "post",
              "author"
            ]
          },
          "id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "id",
          "label"
        ]
      },
      "MediaAssetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sourceName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "bytes": {
            "type": "number"
          },
          "width": {
            "type": "number",
            "nullable": true
          },
          "height": {
            "type": "number",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "quarantined",
              "ready",
              "rejected"
            ]
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          },
          "altText": {
            "type": "string",
            "nullable": true
          },
          "credit": {
            "type": "string",
            "nullable": true
          },
          "rightsNote": {
            "type": "string",
            "nullable": true
          },
          "focalX": {
            "type": "number",
            "nullable": true
          },
          "focalY": {
            "type": "number",
            "nullable": true
          },
          "variants": {
            "description": "Empty until processing finishes; a quarantined asset has no public URL",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaVariantDto"
            }
          },
          "usages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaUsageDto"
            }
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "sourceName",
          "mimeType",
          "bytes",
          "width",
          "height",
          "status",
          "rejectionReason",
          "altText",
          "credit",
          "rightsNote",
          "focalX",
          "focalY",
          "variants",
          "usages",
          "version",
          "createdAt"
        ]
      },
      "RequestUploadDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "maxLength": 255,
            "description": "Original file name, for admin display only"
          },
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          },
          "bytes": {
            "type": "number",
            "minimum": 1,
            "maximum": 10485760
          }
        },
        "required": [
          "fileName",
          "contentType",
          "bytes"
        ]
      },
      "UploadTicketDto": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string"
          },
          "uploadUrl": {
            "type": "string",
            "description": "Short-lived signed PUT URL; the key is server-generated"
          },
          "headers": {
            "type": "object",
            "description": "Headers that must accompany the PUT"
          },
          "expiresInSeconds": {
            "type": "number"
          }
        },
        "required": [
          "assetId",
          "uploadUrl",
          "headers",
          "expiresInSeconds"
        ]
      },
      "CompleteUploadDto": {
        "type": "object",
        "properties": {
          "checksum": {
            "type": "string",
            "description": "SHA-256 of the uploaded bytes, hex; verified when supplied"
          },
          "altText": {
            "type": "object",
            "maxLength": 255,
            "description": "Alt text; required before the asset can be used on a page (SRS MED 003)"
          }
        }
      },
      "UpdateMediaDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number"
          },
          "altText": {
            "type": "string",
            "nullable": true,
            "maxLength": 255
          },
          "credit": {
            "type": "string",
            "nullable": true,
            "maxLength": 255
          },
          "rightsNote": {
            "type": "string",
            "nullable": true,
            "maxLength": 500
          },
          "focalX": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Focal point as a fraction of the width"
          },
          "focalY": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        },
        "required": [
          "expectedVersion"
        ]
      },
      "GalleryEntryDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string"
          },
          "sortOrder": {
            "type": "number"
          },
          "caption": {
            "type": "string",
            "nullable": true
          },
          "alt": {
            "type": "string",
            "description": "Alt text actually used: the override when set, otherwise the asset alt text"
          },
          "isCover": {
            "type": "boolean"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaVariantDto"
            }
          }
        },
        "required": [
          "mediaId",
          "sortOrder",
          "caption",
          "alt",
          "isCover",
          "variants"
        ]
      },
      "GalleryItemDto": {
        "type": "object",
        "properties": {
          "mediaId": {
            "type": "string"
          },
          "caption": {
            "type": "string",
            "nullable": true,
            "maxLength": 255
          },
          "altOverride": {
            "type": "string",
            "nullable": true,
            "maxLength": 255,
            "description": "Alt text for this context; falls back to the asset alt text"
          },
          "isCover": {
            "type": "boolean",
            "description": "Exactly one item may be the cover"
          }
        },
        "required": [
          "mediaId"
        ]
      },
      "SetGalleryDto": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "number"
          },
          "items": {
            "maxItems": 24,
            "description": "Replaces the gallery; order is the array order",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GalleryItemDto"
            }
          }
        },
        "required": [
          "expectedVersion",
          "items"
        ]
      },
...
    }
  }
}